refactor(ui): adjust layout & refactor toolbar
