+++ packages/ui/src/views/app.module.less
@@ -21,12 +21,6 @@
             width: 100%;
         }
 
-        &-main {
-            display: flex;
-            flex: 1;
-            flex-direction: column;
-        }
-
         &-wrapper {
             display: grid;
             grid-template-columns: 1fr auto;
